backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
pandas-ta==0.3.14b0

# Machine Learning
torch==2.1.1
scikit-learn==1.3.2

# APIs
ccxt==4.1.13
websockets==12.0
httpx==0.25.1

# Task Queue
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Documentation
fastapi-cli==0.0.4
python-multipart==0.0.6
